publishing/requirements.txt
ffmpeg
ffmpeg-python
numpy
opencv-python
picamera
python-dotenv